requirements.txt
Flask>=3.0,<4.0
flask-cors
gunicorn
python-dotenv
groq
